Backend/requirements.txt
africastalking==1.2.7
asgiref==3.8.1
certifi==2024.6.2
cgmzscore==3.3.1
charset-normalizer==3.3.2
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.2
django-phone-field==1.8.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
phonenumbers==8.13.39
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.32.3
schema==0.7.7
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.2
